updated Mongo Csharp driver to version 1.3.0.4309

+++ src/log4mongo-net.Tests/MongoDBAppenderTests.cs
@@ -158,7 +158,7 @@ namespace log4net_MongoDB.Tests
             Assert.AreEqual(innerException["message"].AsString, "I'm the inner", "Inner exception message different from expected");
         }
 
-        protected int GetCollectionCount()
+        protected long GetCollectionCount()
         {
             return collection.Count();
         }
